requirement.txt
certifi==2023.5.7
charset-normalizer==3.1.0
dask==2023.3.2
docopt==0.6.2
fonttools==4.39.3
fsspec==2023.4.0
idna==3.4
importlib-metadata==6.3.0
kiwisolver==1.4.4
locket==1.0.0
matplotlib==3.7.1
mplcursors==0.5.2
networkx==3.1
numpy==1.24.2
packaging==23.0
partd==1.3.0
Pillow==9.5.0
psutil==5.9.4
pyparsing==3.0.9
pyperf==2.6.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt5-stubs==5.15.6.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
six==1.16.0
toolz==0.12.0
urllib3==2.0.3
Yaml8==0.1.2
yarg==0.1.9
zipp==3.15.0
pandas==2.2.3